Implement all express routes for Company use case

+++ application/usecases/company/DeleteCompany.ts
@@ -1,9 +1,14 @@
 import { CompanyRepository } from "../../repositories/CompanyRepository";
+import { CompanyNotFound } from "../../../domain/errors/CompanyNotFound";
 
 export class DeleteCompany {
   public constructor(private readonly companyRepository: CompanyRepository) {}
 
   public async execute(id: string) {
+    const company = await this.companyRepository.findById(id);
+    if (!company) {
+      return new CompanyNotFound();
+    }
     await this.companyRepository.delete(id);
   }
 }
